tests/test_inference.py
import os
import pytest
import numpy as np
import torch
from pathlib import Path
from unittest.mock import Mock, patch
from huggingface_hub import HfApi
from dotenv import load_dotenv
import shutil

from inference import StyleBertVITS2, TTSModelManager
from schema import ModelConfig

# .envファイルの読み込み
load_dotenv()

# 環境変数の取得
HF_TOKEN = os.getenv("HUGGING_FACE_TOKEN")
MODELS_CACHE_DIR = Path(os.getenv("MODELS_CACHE_DIR", "models_cache"))

# Hugging Face APIの初期化
hf_api = HfApi(token=HF_TOKEN)

@pytest.fixture
def setup_model_files(tmp_path):
    """モデルファイルのセットアップ"""
    # テスト用のディレクトリ構造を作成
    model_dir = tmp_path / "test_model"
    model_dir.mkdir()
    
    # テスト用のファイルを作成
    config_path = model_dir / "config.json"
    model_path = model_dir / "model.safetensors"
    style_vectors_path = model_dir / "style_vectors.npy"
    
    # ダミーファイルの作成
    config_path.write_text('{"data": {"sampling_rate": 22050}}')
    model_path.write_bytes(b"dummy_model_data")
    np.save(style_vectors_path, np.zeros((1, 10)))
    
    return {
        "model_dir": model_dir,
        "config_path": str(config_path),
        "model_path": str(model_path),
        "style_vectors_path": str(style_vectors_path)
    }

@pytest.fixture
def mock_tts_model(monkeypatch):
    """TTSModelのモック"""
    mock = Mock()
    mock_instance = Mock()
    mock_instance.infer.return_value = (22050, np.zeros(22050, dtype=np.float32))
    mock_instance.hps = Mock()
    mock_instance.hps.data = Mock()
    mock_instance.hps.data.sampling_rate = 22050
    mock_instance.load.return_value = None  # load()メソッドのモック
    mock.return_value = mock_instance
    
    # TTSModelをモックに置き換え
    monkeypatch.setattr("style_bert_vits2.tts_model.TTSModel", mock)
    
    # get_net_g関数もモック
    def mock_get_net_g(*args, **kwargs):
        return Mock()
    monkeypatch.setattr("style_bert_vits2.models.infer.get_net_g", mock_get_net_g)
    
    return mock

def test_huggingface_auth():
    """Hugging Face認証のテスト"""
    assert HF_TOKEN is not None, "HUGGING_FACE_TOKENが設定されていません"
    
    # トークンの有効性を確認
    try:
        hf_api.whoami()
    except Exception as e:
        pytest.fail(f"Hugging Face認証に失敗しました: {str(e)}")

def test_find_latest_model():
    """最新モデルの検索機能のテスト"""
    repo_id = "LiveTaro/uranai-emotion2"
    model_id = "uranai-emotion2"
    
    try:
        # 固定のファイル名を使用
        latest_file = f"{model_id}.safetensors"
        print(f"モデルファイル: {latest_file}")
        
        assert latest_file.endswith(".safetensors")
        
    except Exception as e:
        pytest.fail(f"モデルファイルの検索に失敗しました: {str(e)}")

def test_model_manager():
    """モデルマネージャーの機能テスト"""
    manager = TTSModelManager()
    model_id = "uranai-emotion2"
    
    # テスト用のディレクトリを作成
    test_dir = Path("test_data")
    test_dir.mkdir(exist_ok=True)
    
    # テスト用のファイルを作成
    config_path = test_dir / "config.json"
    model_path = test_dir / f"{model_id}.safetensors"
    style_vectors_path = test_dir / "style_vectors.npy"
    
    config_path.write_text('{"data": {"sampling_rate": 22050}}')
    model_path.write_bytes(b"dummy_model_data")
    np.save(style_vectors_path, np.zeros((1, 10)))
    
    # モデル設定の作成
    config = ModelConfig(
        model_id=model_id,
        name="テストモデル",
        description="テスト用",
        model_path=str(model_path),
        config_path=str(config_path),
        style_vectors_path=str(style_vectors_path),
        speakers={"0": "通常"}
    )
    
    # モデルの追加
    manager.add_model(config, set_as_default=True)
    assert manager.default_model_id == model_id
    
    # モデルの取得
    retrieved_config = manager.get_config(model_id)
    assert retrieved_config.name == "テストモデル"
    
    # モデルの削除
    manager.remove_model(model_id)
    assert manager.default_model_id is None
    
    # テスト用のディレクトリを削除
    shutil.rmtree(test_dir)

@pytest.mark.asyncio
async def test_model_download(tmp_path):
    """モデルのダウンロード機能のテスト"""
    from app import download_model_files
    
    repo_id = "LiveTaro/uranai-emotion2"
    model_id = "uranai-emotion2"
    
    try:
        paths = await download_model_files(model_id, repo_id)
        
        assert "model_file" in paths
        assert "config_file" in paths
        assert "style_vectors_file" in paths
        
        assert os.path.exists(paths["model_file"])
        assert os.path.exists(paths["config_file"])
        assert os.path.exists(paths["style_vectors_file"])
        
    except Exception as e:
        pytest.fail(f"モデルのダウンロードに失敗しました: {str(e)}")

@pytest.mark.skipif(not torch.cuda.is_available(), reason="GPUが必要なテスト")
def test_model_inference(setup_model_files, mock_tts_model):
    """モデルの推論機能のテスト"""
    try:
        model = StyleBertVITS2(
            model_path=setup_model_files["model_path"],
            config_path=setup_model_files["config_path"],
            style_vectors_path=setup_model_files["style_vectors_path"]
        )
    except Exception as e:
        pytest.skip(f"モデルの初期化に失敗しました: {str(e)}")

    # 推論のテスト
    text = "こんにちは"
    try:
        audio = model.infer(
            text=text,
            speaker_id="0"
        )
        
        assert isinstance(audio, np.ndarray)
        assert audio.dtype == np.float32
        assert len(audio.shape) == 1  # モノラル音声
        assert len(audio) > 0  # 音声データが空でないことを確認
        
    except Exception as e:
        pytest.fail(f"推論に失敗しました: {str(e)}")

def test_parameter_validation(mock_tts_model):
    """パラメータバリデーションのテスト"""
    # テスト用のディレクトリを作成
    test_dir = Path("test_data")
    test_dir.mkdir(exist_ok=True)
    
    # テスト用のファイルを作成
    config_path = test_dir / "config.json"
    model_path = test_dir / "uranai-emotion2.safetensors"
    style_vectors_path = test_dir / "style_vectors.npy"
    
    config_path.write_text('{"data": {"sampling_rate": 22050}}')
    model_path.write_bytes(b"dummy_model_data")
    np.save(style_vectors_path, np.zeros((1, 10)))
    
    model = StyleBertVITS2(
        model_path=str(model_path),
        config_path=str(config_path),
        style_vectors_path=str(style_vectors_path)
    )
    
    # 不正なパラメータでのテスト
    with pytest.raises(ValueError):
        model.infer(text="test", speaker_id="-1")  # 不正な話者ID
    
    # テスト用のディレクトリを削除
    shutil.rmtree(test_dir) 