pyproject.toml
[project]
name = "dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=22.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "huggingface-hub>=0.29.3",
    "inflect>=5.3.0",
    "jieba>=0.42.1",
    "ko-pron>=1.3",
    "librosa>=0.8.0",
    "numpy==1.26.4",
    "phonemizer>=3.0.0",
    "pydantic>=2.10.6",
    "pyopenjtalk>=0.3.0",
    "pypinyin>=0.44.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "scipy>=1.7.0",
    "sentencepiece>=0.1.97",
    "setuptools>=76.0.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
    "style-bert-vits2>=2.5.0",
    "sudachidict-core>=20250129",
    "sudachipy>=0.6.10",
    "torch>=2.0.0",
    "transformers>=4.20.0",
    "unidecode>=1.3.0",
    "uvicorn>=0.34.0",
]
